+++ MathUtils.h
@@ -14,7 +14,7 @@ public:
 
     static float periodicmod(float x, float y);
     static float inrange(float x, float min, float max);
-    static float constrain(float value, float min, float max);
+    static double constrain(double value, double min, double max);
     static float deg2rad(float deg);
     static float rad2deg(float rad);
     static float micros2sec(float sec);
